Adding views

+++ common_etl/utils.py
@@ -606,7 +606,12 @@ def create_view_from_query(view_id, view_query):
     view = bigquery.Table(view_id)
 
     if exists_bq_table(view_id):
-        client.delete_table(view_id)
+        if view.table_type == 'VIEW':
+            client.delete_table(view_id)
+        else:
+            has_fatal_error(f"""
+            {view_id} already exists, and refers to a non-view type ({view.table_type}). Cannot create view, exiting.
+            """)
 
     view.view_query = view_query
     view = client.create_table(view)
